Melhorar extendedinfo
